basic commands lists and channels filter
the basic commands are now on the "on message" handler, at least the basic ones are finally complete. It's necessary to implement the functions of each command. There was also a small filter for which type of channel the message came from. The bot only read messages from text channels and not direct message channels of any kind.